+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Workflow DAG
+- Generic plotter for error rates for randomized benchmarks
 
 ### Changed